+++ android/src/main/java/com/brentvatne/react/PlayerEventListener.java
@@ -2,13 +2,13 @@ package com.brentvatne.react;
 
 public interface PlayerEventListener {
     void onLoad(String uriString, String type, boolean isNetwork);
-    void onLoadComplete(int currentPosition, int mVideoDuration);
+    void onLoadComplete(long currentPosition, long mVideoDuration);
     void onPlay();
     void onPause();
     void onStop();
-    void onProgress(int curPos);
-    void onBuffer(int percent, int duration);
-    void onSeek(int msec, int currentPosition);
+    void onProgress(long curPos);
+    void onBuffer(int percent, long duration);
+    void onSeek(long msec, long currentPosition);
     void onError(int what, int extra);
     void onVolume(float leftVolume, float rightVolume);
     void onMute(boolean muted);
